Refactor: revert using typescript in tests
Using typescript in tests seem to be more complicated than it sounds.
Since we need to mock around objects and sometimes use global objects,
it turns out to be very cumbersome. Also configuring it properly is
probably not worth the effort. Therefore, I'm going to revert all
ts files to js in tests.

Also, to separate the test helpers from the src files, I've moved
out the testing folders at the same level with source.