chore: tabify environment configuration
Display various options in a tabified view. Prior to this change,
the custom storage settings were accessible only through a popup
menu in the Environment page. This was not adding visibility to the
feature. Now, it's displayed on the environment config page as a
separate tab. This also allows us to extend the page and more settings.